Cleanup and restructure of repository for rewrite
- Cronex is being rewritten. To make the repository easier to work with,
  the original code has been renamed from cronex to cronex.legacy and
  the unit tests updated accordingly.
- The README and CHANGELOG.deprecated files have been removed; the
  README will need to be rewritten, the contents of CHANGELOG.deprecated
  are irrelevant to the rewrite.
- The setup.py and setup.cfg files have been removed but will be
  introduced once the library is ready to be released.
