--- CHANGELOG.deprecated
@@ -1,82 +1,0 @@
-This file is no longer updated in favor of writing descriptive commit messages,
-but it shall remain for posterity since I do not wish to rewrite the Git
-repository history.
-
-                                      ---
-
-2011.01.16; Sunday
-    Fixed a big that caused one of the tests to fail in certain months of the
-    year. The bug was only in the test, not the trigger code.
-
-2010.12.14; Tuesday
-    Fixed mis-spelling of "annually" and rewrote much of the SYNTAX
-    documentation.
-
-2010.12.07; Tuesday
-    Fixed bug in which wrap-around ranges sometimes skip values.
-
-2010.11.29; Monday
-    Flawed implementation of the day of the week / day of the month changed
-    made on 2010.11.15 corrected and "test_dom_either_or_dow" made more
-    comprehensive.
-
-2010.11.25; Thursday
-    Nicer exception is raised when the number of fields in the CronExpression
-    is too short.
-
-    Works with Python 3.1.
-
-2010.11.21; Sunday
-    CronExpression now validates the syntax of special characters "L", "W",
-    and "#". The unit tests have been changed to reflect this. A new
-    function, "is_special_atom", was created in the process of making this
-    change.
-
-2010.11.17; Wednesay
-    Improvements in the unit tests.
-
-    Moved the original README.markdown to a documentation folder and renamed
-    it to SYNTAX.
-
-    Documentation revamped.
-
-2010.11.16; Tuesday
-    The epoch is now a six element tuple. The last element represents offset
-    from UTC. CronexExrepssion.check_trigger now uses a tuple for the date and
-    accepts and additional option for UTC offset.
-
-    Documentation removed from the module docstring. Method docstrings
-    augmented with a bit more information.
-
-    Exception now raised if "*" is used with other elements in the same field.
-
-    Periodic triggers in the field for days of the week is properly
-    implemented but may not be documented or tested.
-
-    Unit tests should now cover all basic cases; finshed W and # unit tests.
-
-2010.11.15; Monday
-    My implementation's handling the day of the month and day of the week was
-    incorrect:
-
-        $ man 5 crontab
-
-        > Note: The day of a command's execution can be specified by two
-        > fields - day of month, and day of week. If both fields are
-        > restricted (i.e., aren't *), the command will be run when either
-        > field matches the current time. For example, "30 4 1,15 * 5" would
-        > cause a command to be run at 4:30 am on the 1st and 15th of each
-        > month, plus every Friday.
-
-    Unit test added to reflect the above change. I now know why some versions
-    of cron use a "?" in the dom / dow fields.
-
-    Added unit tests for month and day of the week substitution.
-
-    Realized I broke L, W, # checking before the initial commit. Fixed and
-    unit testing added for L in dow and dom fields.
-
-    Renamed some variables.
-
-2010.11.14; Sunday
-    First version is done, unit tested, and documented.
