+++ cronex/legacy/tests.py
@@ -22,7 +22,7 @@ if isinstance(map, type):
 sys.path.insert(0, os.path.join(
     os.path.dirname(os.path.realpath(__file__)), os.pardir))
 
-import cronex
+import cronex.legacy as cronex
 
 
 class test_testedmodule(unittest.TestCase):
